FRONTEND-1890 Consolidate utils folders
