+++ src/utilities/replaceUserCreds.js
